feat/exception handlers (#27)
An additional file exceptions.py is added to override the OpenAI and Cohere Routers within the backend, adding a custom exception handler for common OpenAI API and Cohere API exceptions. The exception handlers catch the raised errors and pass the exception message within the JSONresponse object so that they can be displayed within the UI and provide greater insight on the error being thrown. They also still propagate the error traceback within the console.